requirements.txt
flask==3.0.3
flask-session==0.8.0
flask-cors==5.0.0
flask-limiter==3.12
flask-sqlalchemy==3.1.1
python-dotenv==1.0.1
requests==2.32.3
redis==5.0.8
gunicorn==22.0.0
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
google-api-core
torch>=2.0.0
transformers>=4.30.0
numpy>=1.24.0
scikit-learn>=1.0.0
